requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Deprecated==1.2.18
distro==1.9.0
Django==4.2.16
django-debug-toolbar==4.4.6
django-grappelli==3.0.9
django-tinymce==4.1.0
idna==3.10
mysqlclient==2.2.6
netaddr==1.3.0
pillow==11.0.0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.3.0
wrapt==1.17.2
yookassa==3.5.0
